+++ PhyGen.Application/Authentication/DTOs/Dtos/RegisterDto.cs
@@ -2,8 +2,6 @@
 {
     public class RegisterDto
     {
-        public string UserName { get; set; } = string.Empty;
-        public string Username { get; set; }
         public string Email { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;
         public string FirstName { get; set; } = string.Empty;
