+++ PhyGen.Application/Authentication/DTOs/Responses/AuthenticationResponse.cs
@@ -3,7 +3,6 @@
     public class AuthenticationResponse
     {
         public string Token { get; set; }
-        public string Username { get; set; }
         public string Email { get; set; }
     }
 }
