+++ PhyGen.Application/Authentication/Models/Requests/RegisterRequest.cs
@@ -9,7 +9,6 @@ namespace PhyGen.Application.Systems.Users
 {
     public class RegisterRequest
     {
-        public string UserName { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
