+++ PhyGen.Domain/PhyGen.Domain.csproj
@@ -10,8 +10,4 @@
     <ProjectReference Include="..\PhyGen.Shared\PhyGen.Shared.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.11" />
-  </ItemGroup>
-
 </Project>
