--- PhyGen.Insfrastructure/Identity/AuthService.cs
@@ -1,81 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.Data;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using Microsoft.IdentityModel.Tokens;
-using PhyGen.Application.Authentication.DTOs.Dtos;
-using PhyGen.Application.Authentication.Interface;
-using PhyGen.Application.Authentication.Responses;
-using PhyGen.Domain.Entities;
-using PhyGen.Domain.Exceptions;
-using PhyGen.Shared.Constants;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text;
-
-namespace PhyGen.Insfrastructure.Identity
-{
-    public class AuthService : IAuthService
-    {
-        private readonly UserManager<User> _userManager;
-        private readonly SignInManager<User> _signInManager;
-        private readonly IJwtTokenGenerator _jwtTokenGenerator;
-
-        public AuthService(
-            UserManager<User> userManager,
-            SignInManager<User> signInManager,
-            IConfiguration configuration,
-    IJwtTokenGenerator jwtTokenGenerator)
-        {
-            _userManager = userManager;
-            _signInManager = signInManager;
-            _jwtTokenGenerator = jwtTokenGenerator;
-        }
-
-        public async Task<AuthenticationResponse> RegisterAsync(RegisterDto dto)
-        {
-            var user = new User
-            {
-                FirstName = dto.FirstName,
-                LastName = dto.LastName,
-                PhoneNumber = dto.PhoneNumber,
-                UserName = dto.Username,
-                Email = dto.Email,
-                Role = "User"
-            };
-
-            var result = await _userManager.CreateAsync(user, dto.Password);
-
-            if (!result.Succeeded)
-            {
-                throw new AppException(StatusCode.RegisterFailed);
-            }
-
-            return new AuthenticationResponse
-            {
-                Email = user.Email,
-                Username = user.UserName,
-                Token = _jwtTokenGenerator.GenerateToken(user)
-            };
-        }
-
-        public async Task<AuthenticationResponse> LoginAsync(LoginDto dto)
-        {
-            var user = await _userManager.FindByEmailAsync(dto.Email);
-            if (user == null)
-                throw new AuthException(StatusCode.UserNotFound);
-
-            var result = await _signInManager.CheckPasswordSignInAsync(user, dto.Password, false);
-            if (!result.Succeeded)
-                throw new AuthException(StatusCode.InvalidPassword);
-
-            return new AuthenticationResponse
-            {
-                Email = user.Email,
-                Username = user.UserName,
-                Token = _jwtTokenGenerator.GenerateToken(user)
-            };
-        }
-    }
-
-}
