--- PhyGen.Insfrastructure/Identity/Result.cs
@@ -1,17 +1,0 @@
-﻿
-namespace PhyGen.Insfrastructure.Identity
-{
-    public class Result
-    {
-        public bool Succeeded { get; private set; }
-        public string Error { get; private set; }
-
-        public static Result Success(string token) => new Result { Succeeded = true };
-        public static Result Failure(string error) => new Result { Succeeded = false, Error = error };
-
-      
-    }
-
-   
-
-}
